Lecture 2.5/2.6 - Custom Annotations + Exception Handling

+++ Week 2 - Spring Boot Web/springbootwebtutorial/src/main/java/com/example/springbootwebtutorial/springbootwebtutorial/services/EmployeeService.java
@@ -2,6 +2,7 @@ package com.example.springbootwebtutorial.springbootwebtutorial.services;
 
 import com.example.springbootwebtutorial.springbootwebtutorial.dto.EmployeeDTO;
 import com.example.springbootwebtutorial.springbootwebtutorial.entities.EmployeeEntity;
+import com.example.springbootwebtutorial.springbootwebtutorial.exceptions.ResourceNotFoundException;
 import com.example.springbootwebtutorial.springbootwebtutorial.repositories.EmployeeRepository;
 import org.modelmapper.ModelMapper;
 import org.springframework.data.util.ReflectionUtils;
@@ -10,6 +11,7 @@ import org.springframework.stereotype.Service;
 import java.lang.reflect.Field;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -24,19 +26,22 @@ public class EmployeeService {
         this.employeeRepository = employeeRepository;
     }
 
-    public EmployeeDTO findById(Long id) {
-        EmployeeEntity employeeEntity =  employeeRepository.findById(id).orElse(null);
-       return modelMapper.map(employeeEntity,EmployeeDTO.class);
+    public Optional<EmployeeDTO> getEmployeeById(Long id) {
+       /// EmployeeEntity employeeEntity =  employeeRepository.findById(id).orElse(null);
+       //return modelMapper.map(employeeEntity,EmployeeDTO.class);
+
+       return employeeRepository.findById(id)
+               .map(employeeEntity -> modelMapper.map(employeeEntity,EmployeeDTO.class));
     }
 
-    public List<EmployeeDTO> findAll() {
+    public List<EmployeeDTO> getAllEmployees() {
         List<EmployeeEntity> employeeEntities = employeeRepository.findAll();
         return employeeEntities.stream()
                 .map(employeeEntity -> modelMapper.map(employeeEntity,EmployeeDTO.class))
                 .collect(Collectors.toList());
     }
 
-    public EmployeeDTO save(EmployeeDTO inputEmployee) {
+    public EmployeeDTO createNewEmployee(EmployeeDTO inputEmployee) {
         EmployeeEntity saveEmployee = modelMapper.map(inputEmployee,EmployeeEntity.class);
         return modelMapper.map(employeeRepository.save(saveEmployee),EmployeeDTO.class);
     }
@@ -44,6 +49,8 @@ public class EmployeeService {
 
 
     public EmployeeDTO updateEmployeeById(EmployeeDTO employeeDTO, Long employeeId) {
+        boolean exists = ifExistsByEmployeeID(employeeId);
+        if(!exists) throw new ResourceNotFoundException("Employee not found for Id :"+ employeeId);
         EmployeeEntity updateEmployee = modelMapper.map(employeeDTO,EmployeeEntity.class);
         updateEmployee.setId(employeeId);
 
@@ -53,8 +60,12 @@ public class EmployeeService {
     public boolean ifExistsByEmployeeID(Long employeeId){
          return employeeRepository.existsById(employeeId);
     }
-    public void deleteEmployeeById(Long employeeId) {
-            employeeRepository.deleteById(employeeId);
+    public boolean deleteEmployeeById(Long employeeId) {
+        boolean exists = ifExistsByEmployeeID(employeeId);
+        if(!exists) return false;
+        employeeRepository.deleteById(employeeId);
+        return true;
+
     }
 
     public EmployeeDTO updatePartialEmployeeById(Long employeeId, Map<String, Object> updates) {
